+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vastly/utils
 
+## 0.4.0
+
+### Minor Changes
+
+- Creates Asset service, reusable AssumeRole utility, upload lambda functioning
+
 ## 0.3.2
 
 ### Patch Changes