+++ services/deploy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # deploy-service
 
+## 0.2.1
+
+### Patch Changes
+
+- Creates Asset service, reusable AssumeRole utility, upload lambda functioning
+- Updated dependencies
+  - @vastly/utils@0.4.0
+
 ## 0.2.0
 
 ### Minor Changes